# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API сервиса iAmOkey",
        "description": "Документация сервиса.",
        "contact": {
            "email": "nfs2025@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/patient/{id}": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "Пациент",
                "description": "Возвращается объект пациента",
                "operationId": "6c5984c20f4c599317890ac2892c2862",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Patient"
                ],
                "summary": "Удалить пациента",
                "description": "Удалить пациента",
                "operationId": "d815d9eab862308ce36b76a063b6a93f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пациента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Patient"
                ],
                "summary": "Обновить пациента",
                "description": "Обновить пациента, возвращается объект пациента",
                "operationId": "4a2b4762c3876a717440dab68e38811c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пациента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Patient"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/patients/": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "Пациенты",
                "description": "Возвращается объект пациентов",
                "parameters": [
                    {
                        "name": "name_sort",
                        "in": "path",
                        "description": "Сортировка по имени asc/desc",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "who_sort",
                        "in": "path",
                        "description": "Сортировка по кем приходится asc/desc",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_edit",
                        "in": "path",
                        "description": "Сортировка по последней дате изменения asc/desc",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Сортировка по имени пациента",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/patient/": {
            "put": {
                "tags": [
                    "Patient"
                ],
                "summary": "Создать нового пациента",
                "description": "Создать нового пациента, возвращается объект пациента",
                "operationId": "496e444d68432a4c19d3ab0a4b5d525b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Patient"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/setting/": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Цена подписки",
                "description": "Возвращается цена подписки",
                "operationId": "8f0a647a82e3e274e5d3d0c1ff322232",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Setting"
                ],
                "summary": "Изменить цену подписки",
                "description": "Изменить цену подписки",
                "operationId": "dd58049829494a4826a9887ac8fbe566",
                "parameters": [
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Цена",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Patient": {
                "title": "Patient",
                "description": "Patient model",
                "properties": {
                    "user_id": {
                        "title": "User_id",
                        "description": "ID пользователя",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "ФИО пациента",
                        "type": "integer",
                        "example": "Иванов Иван Иванович"
                    },
                    "birth": {
                        "title": "Birth",
                        "description": "Дата рождения",
                        "type": "string",
                        "example": "06.08.1960"
                    },
                    "gender_id": {
                        "title": "Gender_id",
                        "description": "Пол",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "growth": {
                        "title": "Growth",
                        "description": "Рост",
                        "type": "integer",
                        "format": "int64",
                        "example": 184
                    },
                    "wieght": {
                        "title": "Wieght",
                        "description": "Вес",
                        "type": "integer",
                        "format": "int64",
                        "example": 66
                    },
                    "stride_lenth": {
                        "title": "Stride_lenth",
                        "description": "Длина шага",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "subscription": {
                        "title": "subscription",
                        "description": "Дата подписки",
                        "type": "integer",
                        "example": null
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Patient"
                }
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}